\name{medals}
\alias{medals}
\docType{data}
\title{
Olympic medals dataset
}
\description{
Medals won and population size for each country in the 2016 Rio Olympics
}
\usage{data("medals")}
\format{
  A data frame with 60 observations on the following 7 variables.
  \describe{
    \item{\code{code}}{Country code}
    \item{\code{gold}}{Number of gold medals won}
    \item{\code{silver}}{Number of silver medals won}
    \item{\code{bronze}}{Number of bronze medals wonr}
    \item{\code{total}}{Total number of medals won}
    \item{\code{country}}{Country name}
    \item{\code{population}}{Population size}
  }
}

\source{
Medal table obtained from https://www.rio2016.com/en/medal-count-country [Accessed on 6 October 2016, but this link is no longer active.]

Population sizes obtained from http://data.worldbank.org/indicator/SP.POP.TOTL?end=2015&start=2000 [Accessed on 6 October 2016]}


\keyword{datasets}
